Parameterize GraphQL retries and extract helpers
