+++ docs/vk-design.md
@@ -130,10 +130,13 @@ classDiagram
 ## GraphQL Error Handling
 
 GraphQL requests are retried when a network error occurs or the response lacks
-data. The client waits `200ms * 2^attempt` plus up to `200ms` of random jitter
-between tries, attempting each query at most five times. Because `run_query`
-only returns after a full page has been fetched, `paginate` never appends
-partial results, preserving order and avoiding duplicates.
+data. Retry behaviour is configurable through `RetryConfig`, covering the
+number of attempts, base delay, and jitter fraction. By default the client
+attempts a query up to five times, waiting `200ms * 2^attempt` plus up to the
+same backoff again of random jitter, scaling jitter with the backoff so
+concurrent callers spread out as delays grow. Because `run_query` only returns
+after a full page has been fetched, `paginate` never appends partial results,
+preserving order and avoiding duplicates.
 
 The diagram below illustrates how deserialisation errors surface the JSON path
 and a response snippet, helping developers quickly locate schema mismatches.
